+++ src/window.rs
@@ -701,6 +701,17 @@ impl Window {
         self.window.is_maximized()
     }
 
+    /// Gets the window's current resizable state.
+    ///
+    /// ## Platform-specific
+    ///
+    /// - **Wayland / X11:** Not implemented.
+    /// - **iOS / Android / Web:** Unsupported.
+    #[inline]
+    pub fn is_resizable(&self) -> bool {
+        self.window.is_resizable()
+    }
+
     /// Sets the window to fullscreen or back.
     ///
     /// ## Platform-specific